Refactor: Movig files in Store/type folders
